+++ packages/brick_offline_first_with_rest_build/test/offline_first_generator/test_rest_config_endpoint.dart
@@ -49,13 +49,13 @@ class RestConfigEndpointAdapter
   final String fromKey = null;
   final String toKey = null;
   final Map<String, Map<String, dynamic>> fieldsToSqliteColumns = {
-    'primaryKey': SqliteColumnDefinition(
+    'primaryKey': RuntimeSqliteColumnDefinition(
       association: false,
       columnName: '_brick_id',
       iterable: false,
       type: int,
     ),
-    'someField': SqliteColumnDefinition(
+    'someField': RuntimeSqliteColumnDefinition(
       association: false,
       columnName: 'some_field',
       iterable: false,