+++ packages/brick_sqlite/lib/sqlite.dart
@@ -1,4 +1,4 @@
-export 'src/sqlite/sqlite_adapter.dart';
-export 'src/sqlite/sqlite_column_definition.dart';
-export 'src/sqlite/sqlite_model_dictionary.dart';
-export 'src/sqlite/sqlite_provider.dart';
+export 'src/sqlite_adapter.dart';
+export 'src/runtime_sqlite_column_definition.dart';
+export 'src/sqlite_model_dictionary.dart';
+export 'src/sqlite_provider.dart';