+++ packages/brick_sqlite/test/__mocks__.dart
@@ -23,7 +23,7 @@ const _demoModelMigrationCommands = [
       onDeleteSetDefault: false),
   InsertColumn('complex_field_name', Column.varchar, onTable: 'DemoModel'),
   InsertColumn('last_name', Column.varchar, onTable: 'DemoModel'),
-  InsertColumn('name', Column.varchar, onTable: 'DemoModel'),
+  InsertColumn('full_name', Column.varchar, onTable: 'DemoModel'),
   InsertColumn('simple_bool', Column.boolean, onTable: 'DemoModel'),
   CreateIndex(
       columns: ['l_DemoModel_brick_id', 'f_DemoModelAssoc_brick_id'],