+++ packages/brick_sqlite_generators/test/sqlite_model_serdes_generator/test_field_with_type_argument.dart
@@ -29,13 +29,13 @@ class FieldWithTypeArgumentAdapter
   FieldWithTypeArgumentAdapter();
 
   final Map<String, Map<String, dynamic>> fieldsToSqliteColumns = {
-    'primaryKey': SqliteColumnDefinition(
+    'primaryKey': RuntimeSqliteColumnDefinition(
       association: false,
       columnName: '_brick_id',
       iterable: false,
       type: int,
     ),
-    'someField': SqliteColumnDefinition(
+    'someField': RuntimeSqliteColumnDefinition(
       association: false,
       columnName: 'some_field',
       iterable: false,