+++ packages/brick_sqlite_generators/test/sqlite_model_serdes_generator/test_sqlite_unique.dart
@@ -23,13 +23,13 @@ class SqliteUniqueAdapter extends SqliteAdapter<SqliteUnique> {
   SqliteUniqueAdapter();
 
   final Map<String, Map<String, dynamic>> fieldsToSqliteColumns = {
-    'primaryKey': SqliteColumnDefinition(
+    'primaryKey': RuntimeSqliteColumnDefinition(
       association: false,
       columnName: '_brick_id',
       iterable: false,
       type: int,
     ),
-    'someField': SqliteColumnDefinition(
+    'someField': RuntimeSqliteColumnDefinition(
       association: false,
       columnName: 'some_field',
       iterable: false,